refactor: Use single controller for edit
